01 - Reading List: Refactored filter feature.
