+++ pruebas/01-reading-list/iivvaannxx/src/features/filter/index.ts
@@ -1,4 +1,4 @@
 export { default as FilterMenu } from './components/menu.svelte'
 
-export { filter } from './feature'
-export { filters } from './store'
+export { filter } from './lib/utils'
+export { filters } from './state/store'
